requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.5
httpx==0.24.1
transformers==4.35.2
torch==2.1.1
pillow==10.1.0
numpy==1.26.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
aiohttp==3.9.1
opencv-python==4.8.1.78
dlib==19.24.0
face-recognition==1.3.0
scikit-learn==1.3.2
sentencepiece==0.1.99
protobuf==4.25.1
lxml==4.9.3
html5lib==1.1
tenacity==8.2.3 